rename folder `util` to `app`
