+++ src/components/app/SelectLocale.vue
